Fix readthedocs documentation generation
readthedocs make the assumption that the binary sphinx-build
will be located in the venv binary directory.

readthedocs logs demonstrate that this binary is actually installed
in <venv root dir>/lib/launch_ros_sandbox.

While the reason for this behavior is unclear, this changes work
around this issue creating a symbolic link to sphinx-build in the
bin directory to ensure that sphinx-build will find the binary
as expected.

To do so, a fake Python package is introduced in
doc/sphinx_build_symbolink. This fake package is added to
.readthedocs.yml so that its setup.py is executed before
documentation generation. Using a fake package ensures that no
production code is touched by this workaround.

This commit also fixes the .readthedoc.yml file by:
 - fixing the path to conf.py which was incorrect
 - removing commented out configuration options

Fix #22

Signed-off-by: Thomas Moulard <tmoulard@amazon.com>