Fix flake8 issue
Build is currently failing because I rephrased some typing
annotations as comments, which flake8 is not parsing.
As a result, the build is failing because flake8 is complaining
that some types are not used. To fix this issue, this PR
is importing modules instead of types directly.

```
=================================== FAILURES ===================================
[31m[1m_________________________________ test_flake8 __________________________________[0m
[1m[31mtest/test_flake8.py[0m:23: in test_flake8
[1m    assert rc == 0, 'Found errors'[0m
[1m[31mE   AssertionError: Found errors[0m
[1m[31mE   assert 1 == 0[0m
----------------------------- Captured stdout call -----------------------------

./launch_ros_sandbox/actions/load_docker_nodes.py:23:1: F401 'asyncio.Task' imported but unused
from asyncio import CancelledError, Future, Task
^

1     F401 'asyncio.Task' imported but unused

32 files checked
1 errors

'F'-type errors: 1
```

Travis log with the current failure:
https://api.travis-ci.org/v3/job/594750108/log.txt

Signed-off-by: Thomas Moulard <tmoulard@amazon.com>