Remove unnecessary argument from workflow
The "yml" parameter has been removed from the codecov/codecov-action as
documented on the repo:

https://github.com/codecov/codecov-action/tree/v1.0.12

Update: We've removed the yml parameter with the latest release of this
action. Please put your custom codecov yaml file at the root of the repo
because other locations will no longer be supported in the future.

Signed-off-by: Thomas Moulard <tmoulard@amazon.com>